asv pc app: add core hpm command and change default value
This patch add core hpm command and change default value.
( version : 0.11.1 )

+++ apps/lvcc_pc/ASVCommandLib/asv_command.c
@@ -130,7 +130,7 @@ ASV_RESULT MakeCommandString( char *outBuf, int32_t outSize, ASV_COMMAND cmd, AS
 double NXatof(char *str)
 {
 	int len = 0, n=0, i=0;
-	float f=1.0, val=0.0;
+	double f=1.0, val=0.0;
 
 	while(str[len])len++;
